Find the background and make a mask
Currently this script blurs the image, sets a heavy gamma value (~10) and finds the only contours left (usually just the seal). It then builds a polygon of the largest shape. 

It also can find the second largest in the event that the seal bisects the image such that there's an upper background and a lower background. Extracting this polygon and making a mask of it has been a challenge (but likely has a very simple solution). 

Seals on rocky backgrounds still not great.

To do:
If/else: If the first and/or last row of the np.array are light, run as normal. If darker, invert image and run.
Find a way to see if the background is contiguous or broken up.