More stable, works well with most images
Now displays two segments where applicable, capturing more of the background.  Still need to go from the contours to a masked array of NaNs.  

I added a histogram equalization step to help find the seal outline, and it helps for most seals, but ruins others. One image's segment is the whole image minus a tiny spot in the corner. Need to figure out how to make some images get the equalization and others not. 

Also still struggling with the contour -> mask -> patches finding  pipeline.